Split out season/episode data sources
